Refactor test suite for large group performance metrics
