+++ .github/workflows/TS_Large.yml
@@ -38,7 +38,7 @@ jobs:
       - name: Run tests with retry
         id: test
         continue-on-error: true
-        run: ./scripts/run-test.sh TS_Large 1
+        run: ./scripts/run-test.sh large 1
       - name: Send Slack notification
         if: always()
         env:
