+++ deprecated/README.md
@@ -1,0 +1,1 @@
+# 🪦 Where scripts come to die
