+++ deprecated/bot-cli.ts
