+++ deprecated/generate-inboxes.ts
