+++ deprecated/generate-installations.ts
