+++ deprecated/packages.sh
