+++ deprecated/test-slack.sh
