+++ deprecated/ts-large.ts
@@ -1,6 +1,7 @@
 import { loadEnv } from "@helpers/client";
 import generatedInboxes from "@helpers/generated-inboxes.json";
 import { logError } from "@helpers/logger";
+import "@helpers/streams";
 import {
   verifyAddMembersStream,
   verifyGroupUpdateStream,
@@ -109,7 +110,7 @@ describe(testName, async () => {
         try {
           console.log("Testing conversation stream with new DM creation");
 
-          // Use the dedicated conversation stream verification helper
+          // Use the dedicated conversation str eam verification helper
           const verifyResult = await verifyAddMembersStream(
             newGroup as Group,
             workers.getWorkers(),
