+++ package.json
@@ -18,6 +18,7 @@
     "functional": "npx vitest run ./functional/*.test.ts --pool=forks --fileParallelism=false",
     "gen:keys": "tsx scripts/generate-keys.ts",
     "gm": "cd bots/gm-bot && tsx src/index.ts",
+    "large": "npx vitest run ./suites/TS_Large/*.test.ts --pool=forks --fileParallelism=false",
     "lint": "eslint .",
     "monitor:dev": "httpstat https://grpc.dev.xmtp.network:443",
     "record": "npx playwright codegen 'https://xmtp.chat/'",
