+++ suites/TS_Large/ts_large_messages.test.ts
@@ -1,0 +1,92 @@
+import { loadEnv } from "@helpers/client";
+import { logError } from "@helpers/logger";
+import { verifyMessageStream } from "@helpers/streams";
+import { setupTestLifecycle } from "@helpers/vitest";
+import { typeofStream } from "@workers/main";
+import { getWorkers, type WorkerManager } from "@workers/manager";
+import { type Conversation } from "@xmtp/node-sdk";
+import { afterAll, describe, expect, it } from "vitest";
+import { saveLog, ts_large_createGroup } from "./helpers";
+
+const testName = "ts_large_messages";
+loadEnv(testName);
+
+describe(testName, async () => {
+  const workersCount = parseInt(process.env.TS_LARGE_WORKER_COUNT ?? "5");
+  const batchSize = parseInt(process.env.TS_LARGE_BATCH_SIZE ?? "50");
+  const total = parseInt(process.env.TS_LARGE_TOTAL ?? "100");
+  const steamsToTest = typeofStream.Message;
+  let workers: WorkerManager;
+  let start: number;
+
+  let testStart: number;
+  let newGroup: Conversation;
+
+  // Hold timing metrics per group size
+  interface SummaryEntry {
+    groupSize: number;
+    messageStreamTimeMs?: number;
+    groupUpdatedStreamTimeMs?: number;
+    conversationStreamTimeMs?: number;
+    syncTimeMs?: number;
+    createTimeMs?: number;
+  }
+
+  const summaryMap: Record<number, SummaryEntry> = {};
+
+  workers = await getWorkers(workersCount, testName, steamsToTest);
+
+  setupTestLifecycle({
+    expect,
+    workers,
+    testName,
+    getStart: () => start,
+    setStart: (v) => {
+      start = v;
+    },
+    getTestStart: () => testStart,
+    setTestStart: (v) => {
+      testStart = v;
+    },
+  });
+
+  for (let i = batchSize; i <= total; i += batchSize) {
+    it(`receiveLargeGroupMessage-${i}: should create a group and measure all streams`, async () => {
+      try {
+        newGroup = await ts_large_createGroup(workers, i, true);
+
+        const verifyResult = await verifyMessageStream(
+          newGroup,
+          workers.getWorkers(),
+          1,
+          "gm",
+          () => {
+            start = performance.now();
+          },
+        );
+
+        const streamTimeMs = performance.now() - start;
+        console.log(
+          `Message stream for ${i} participants took ${streamTimeMs.toFixed(2)}ms`,
+        );
+
+        // Save metrics
+        summaryMap[i] = {
+          ...(summaryMap[i] ?? { groupSize: i }),
+          messageStreamTimeMs: streamTimeMs,
+        };
+
+        expect(verifyResult.allReceived).toBe(true);
+      } catch (e) {
+        logError(e, expect.getState().currentTestName);
+        throw e;
+      }
+    });
+  }
+
+  // Aft
+  // After all tests have run, output a concise summary of all timings per group size
+  afterAll(() => {
+    saveLog(summaryMap);
+  });
+});
